Got the unit tests passing, but why are they soo much faster? Mystery

+++ src/bufr/BufrParser/Query/QueryRunner.cpp
@@ -216,7 +216,7 @@ namespace bufr {
             }
 
             if (query.path.back()->index > 0 &&
-                query.path.back()->index <= gsl::narrow<int>(targetNodes.size()))
+                query.path.back()->index <= targetNodes.size())
             {
                 targetNodes = {targetNodes[query.path.back()->index - 1]};
             }
